backend/requirements.txt
django
djangorestframework
django-oauth-toolkit
django-cors-headers
drf_social_oauth2